Move the define to the Qt4 support folder
This is not necessary on Linux when building with gcc. However, I will not put
this into the win32 { ... } branch that can be found in that install-helper.pri
already because I have not tested this for cross-compilation with mingw on Linux
to Windows. I am not sure about Mac either. So, even though this is not
necessary on Linux, it is safer to but into the "common" not to cause breakages
for now.

This is for sure not necessary for Qt5 as none of the other modules use such
defines that I have checked, at least.

Change-Id: Ic4d3f0a7d5687e1a91271f09574fe7b99738cb59
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>
Reviewed-by: Sergey Belyashov <Sergey.Belyashov@gmail.com>