Prefer the library built in the project tree for the examples and tests
This is necessary when the user has a previous QtSerialPort version installed
because the examples and tests will pick up the old installed version. That can
cause breakages for the "one shot" build in case of API breaks or behavior
changes.

For consistency, and more importantly for proper explicit usage without raw
INCLUDEPATH usage, QMAKE_INCDIR is introduced in place.

qtAddLibrary has to be moved to the beginning not to prepend the QtSerialPort
include later, and hence to avoid the issue for the example and test builds.

This change also removes a left-over line for Mac as it has just been tested
that the line is unnecessary.

Change-Id: If391b4398a323e0a3eb285bd79d98ec1dd410381
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>
Reviewed-by: Sergey Belyashov <Sergey.Belyashov@gmail.com>