Remove description for QSerialPort::close() method
Description of the method close() does not reflect the reality
in terms of flush data before closing, at current moment data
is not flushes and device closes immediately.

Thus, it is wiser to remove the description to the functionality
that does not exist at the moment. Moreover, that the question
of whether to do or not do flush, is to be considered later.

Change-Id: Ifb286c0c4a608b8dfea8b0380ce92bf3df9f3de3
Reviewed-by: Laszlo Papp <lpapp@kde.org>