Windows: Add missed error code ResourceError processing
Previously, when pulled from the system the USB Bluetooth dongle (which
is associated with the opened serial port) don't cause emitting of signal
error() with code ResourceError. This reproduced only in case if through
already open the serial port was called the read operation, in this case
read operation returns fail with the system error code ERROR_DEVICE_REMOVED.

Now, this problem has been fixed and handled correctly by at least on
Windows XP, Vista, 7 and 8 (x32) with the Standard MS Bluetooth stack.

Seems, that it is common behavior for all the USB Bluetooth devices with
use the Standard MS Bluetooth Stack.

Change-Id: Ic19c850360d27a2729f72b438e1e185db09eeb97
Reviewed-by: Sergey Belyashov <Sergey.Belyashov@gmail.com>
Reviewed-by: Laszlo Papp <lpapp@kde.org>