Fix crash the Terminal example when trying to select the custom baud rate
Attempting to choose from QComboBox the custom baud rate leads to
crash because the null pointer is returned to the QLineEdit and call
the method for setup the validator by null pointer.

Only editable combo boxes have a line edit, so trying to set a
non-editable is dereferencing a null ptr.

Change-Id: Ia9aaef64ef908b1b2d95090a203dc223d12f0b76
Reviewed-by: Sergey Belyashov <Sergey.Belyashov@gmail.com>
Reviewed-by: Laszlo Papp <lpapp@kde.org>