Make a good use of the new "override" C++11 specifier when available
The usage of the specifier helps to catch mistakes in general and also
explicitly shows the intention.

Furthermore, leave the "virtual" specifier out as per Qt Coding Style:
http://qt-project.org/wiki/Qt_Coding_Style#fe675674b70cdaca505f3cff0244fbae

It does not make much sense in general with the override specifier either
anyhow.

Note: an empty define is still necessary as long as QtSerialPort supports Qt4
which is the current situation.

Change-Id: Iedf01071e6303d88305c140f4a767bb43059f593
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>
Reviewed-by: Olivier Goffart <ogoffart@woboq.com>