Add local echo support for the terminal example
Earlier, by default when put a characters for the transfer, this
characters is always displayed on the console, i.e. can be said
that has always worked the local echo mode.

This led to the problem with duplication of input characters on the
console, in case if the remote device (e.g. a modem) was enabled the
echo mode by default. For example, when typing the command "AT", the
console displays "AATT".

To fix this problem, to the settings dialog was added the checkbox
"Local echo" in which can turn off the display of input characters
to the console.

Change-Id: Iebc9384bab97e313a954a8f8c00ea22129913280
Reviewed-by: Laszlo Papp <lpapp@kde.org>
Reviewed-by: Sergey Belyashov <Sergey.Belyashov@gmail.com>