Mac: Fixed the description() method for Bluetooth devices
At the moment, the description() method returns empty string, because
for Bluetooth devices does not exist the properties with the names
kIOPropertyProductNameKey and kUSBProductString into dictionary.

To obtain description information for the Bluetooth serial devices
can to use a separate property called "BTName". This property can also
be seen with the utility IORegistryExplorer.

In the current implementation, the name "BTName" is hard-coded in the
source code because I did not find any of its definition in the header
files XCode, in the Apple's developer site and in general, on the hard
drive. Thus, in the future it is possible to replace the hard-code to
a constant, if it exists.

Tested on MacOSX 10.6.8

Task-number: QTPLAYGROUND-25
Change-Id: I227aa2a1429660287ea83c1b142d2af3a11fd086
Reviewed-by: Sergey Belyashov <Sergey.Belyashov@gmail.com>
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>