Fix of errors of compilation on WinCE platform
Some previous commits entered regressions in compile process,
namely:

* sha1: 544f927c3724344d01d45d19151d236c143f8c83

After implementation of a new concept of internal architecture
of library, is entered regression for the indefinite eventMask
variable in the waitForReadOrWrite() method.

Solution - to define a variable into this method.

* sha1: 3ea635d74d4035c9036c2b7208b162620f51aa8b

Entered the name space regression for the class CommEventNotifier.

Solution - not to use a name space for this class.

Tested for target:

"-platform win32-msvc2005 -xplatform wincewm50pocket-msvc2005"

Task-number: QTPLAYGROUND-13
Change-Id: I9dbcc0023f06c1b0e6d7b1b670828650382ab785
Reviewed-by: Laszlo Papp <lpapp@kde.org>