Windows: Fix waitAnyEvent() method
Wrong comparison between unsigned variable waitResult and signed size of vector.
The waitResult variable was interpreted as signed data type, that led to an
wrong result of comparison (e.g. when waitResult = WAIT_FAILED (0xFFFFFFFF)).

The solution - convert the size of the array to an unsigned data type and
leave waitResult without conversion.

Thanks to "boo kan, ong".

Task-number: QTBUG-32017
Change-Id: Iff25304d3d0345f47e3edc363a999a7b979cb1bb
Reviewed-by: Sergey Belyashov <Sergey.Belyashov@gmail.com>
Reviewed-by: Laszlo Papp <lpapp@kde.org>