Revert "Add property to set exclusivity"
This reverts commit 4c67500e5d14e561e37b4bdc056f059d3eefdfa4.

"dev: unfrozen branch, containing alpha-quality code that is ready to go into
beta testing at any time"

... from: http://qt-project.org/wiki/Branch-Guidelines

Please respect the basic principles. In doubts, do not approve changes yourself,
but ask for help.

Besides the illegit nature of this change, it has fundamental documentation
issues as well. It is summer, and this can wait until 5.2, so please do /not/
rush with approvals.

Change-Id: Ic2763394035c776088cb28f9c04086920cedb3a2
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>