Do not set an error for timeout as there is no timeout error handled yet
This will need a proper extension later with a dedicated timeout error for 5.2
which will follow this change.

See the following QIODevice subclasses for details to get a rough idea:

http://doc-snapshot.qt-project.org/qt5-stable/qtcore/qprocess.html#ProcessError-enum
http://qt-project.org/doc/qt-5.0/qtcore/qfiledevice.html#FileError-enum
http://doc-snapshot.qt-project.org/qt5-stable/qtnetwork/qabstractsocket.html#SocketError-enum

This will also mean, we need to duplicate the qt_safe_select and
qt_core_unix.cpp etc files as qt_safe_select is exported, but only from a
private header. Anyway, more explain to come for that change.

Task-number: QTBUG-32016
Change-Id: If10cd7b5575636f0f7ee3368258e80bd4766ff3a
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>