Add function to get serial port handle
Sometimes the Qt functions are just not enough, and for this reason Qt
provides functions like QFileDevice::handle(), QAbstractSocket::
socketDescriptor(), QWidget::winId(), and QProcess::pid(). QSerialPort
should do the same. In my case, I find that QSerialPort is easy to use
and works nicely with the rest of Qt, but my application needs to set
TIOCNXCL right after the serial port is opened.

Change-Id: Ie740e478b5825fa2e699ead079d78c514a0caef4
Reviewed-by: Laszlo Papp <lpapp@kde.org>
Reviewed-by: Mika Korhonen <mika.j.korhonen@gmail.com>
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>