Fix the compilation for Android with Qt 5.
1) The termios struct is basically copied and pasted from the documentation:
http://linux.die.net/man/3/termios

Since, it is just an interface without complicated methods, it is not
copyrightable, so it is safe enough.

2) This is also the same situation with the two defines we need for the kernel
interface. They are just interface numbers, and not something that
copyrightable. This has also been double checked and discussed with Lars in
email.

3) The ::tcdrain convenience method cannot be used on Android, so the fallback
"raw" ioctl syscall is used with the relevant parameters, e.g. TCSBREAK.

http://linux.die.net/man/3/tcdrain

5) cfsetspeed seems to be a BSD 4.4 extension based on the termios(3) man page:

http://linux.die.net/man/3/termios

"cfsetspeed() is a 4.4BSD extension.  It takes the same arguments as
cfsetispeed(), and sets both input and output speed."

... hence this is replaced with separate calls for input and output without a
specific Android "ifdef" in the code. It is actually also more consistent with
the setcfispeed and setcfospeed just right above that snippet.

Task-number: QTBUG-33139

Change-Id: If0abb8cec9e900fa2b6fc24df938cb778a344d55
Reviewed-by: Sergey Belyashov <Sergey.Belyashov@gmail.com>
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>