Revert "Android: Fix compilation on Necessitas Qt SDK v 1.0.0"
This reverts commit b80b2eff9d4f0ce86ac168bc758580d6587d0b87.

This change is broken, and was integrated by misleading concepts. This has to be
reverted, and the proper fix will come from stable when the merge happens.
Otherwise, the merge will be painful for the people doing it without expertise
and external communication.

You can find the change on the following URL with the proper design to address
the issue:

https://codereview.qt-project.org/#change,64320

Change-Id: Ib77f2941a7c24a7b0597a4882e7f35b88b7802ee
Reviewed-by: Sergey Belyashov <Sergey.Belyashov@gmail.com>
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>