Do not set the sysfs availability to false on Linux
It is unnecessary since it will be evaluated again. As a side note, it can be a
static variable since it is set to false which cannot cause thread-safety
issues. Even more, these variables can get the "const" qualifier as they are not
being changed.

Tested on Linux with Qt 4 and then Qt 5.

Change-Id: I5d9a154b703d794842964c974f44dc75caffec59
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>
Reviewed-by: Sergey Belyashov <Sergey.Belyashov@gmail.com>