Add a new error value called "NotOpenError"
This error message is used when the serial port parameter setters are called
even though the serial port is not open, and it should be.

The details are very similar to the previously introduced TimeoutError, as in:

1) The "\since 5.2" notion has to be added manually as qdoc has no support for
enumeration values in that regard.

2) The value itself has to be added at the end of the enumeration not to break
the binary compatibility unnecessarily. It is fine as long as the error
documentation for the end users can be kept in the order we wish to do.

Each method documentation already contains the information that the error is set
properly when that occurs, so this change can be considered as a bug fix rather
than a new feature. It is a behavior change of course just as any bug fix.

The previous operation was wrong to go ahead down to the syscalls even though
the serial port was not open. Thereby, that was a bit pointless.

Unfortunately, there is no similar error introduced for the existing QIODevice
subclasses just yet, but this term is at least inline with the open mode of
QIODevice which is "NotOpen".

For sure, when the parameter setters are called by the end user without the
device being open results false return value as per method documentation.

This change was tested on Linux with Qt 4 and then 5.

Change-Id: I1bc7309e34bdf59793f1de510866dae1bb48b539
Reviewed-by: Sergey Belyashov <Sergey.Belyashov@gmail.com>
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>