Return in certain methods if the device is not open
This is necessary for a couple of methods because they require the port being
open for being functional. Hence, if the port is not open, they now return
properly on their own, and set the "NotOpenError" error code .

This has been (only build!) tested on Linux (Arch) with Qt 4 and then 5.

Change-Id: Ifc77bea581dd1832e188fa7562c5c1b15c0532e3
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>