Eliminate the ifdef in the code for imported headers (qt4support)
As I previously wrote on Gerrit, this idea of resembling the include hiearchy
inside qt4support has two advantages:

1) The includes are well-separated for Qt 4 that are imported.

2) The code is less complex due to the non-existence of ifdefs for major Qt
variants.

The only disadvantage is that the build system got one additional line, but the
benefits outweigh that.

This has been (only build!) tested on Linux (Arch) with Qt 4 and then 5.

Change-Id: Ic33632cd0e569693a65af094479dac0a02b1a7f7
Reviewed-by: Sergey Belyashov <Sergey.Belyashov@gmail.com>
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>