Fix the broken example qdoc keyword value
The "cwritersync" example has been introduced in commit
9f8f2d6dc6f38b54a91f4a3d7ebf4285891005d3, and the "creadersync" example has been
introduced in commit ccdf42d9d10a32ccdd4a12c643130ffdea498552.

Initially, the "cwritersync" example was called "cwriter", but then it got
renamed to reflect the functionality more explicitly based on its name.

Having that renaming done, I had to sync up the writer and reader terms
everywhere, but this instance has been accidentally left out. It is not
intentional to have this name, so this is a bug.

This is a somewhat critical bugfix because it may trigger the example
unavailable from the main example page. Thereby, a useful example could become
useless for the end users.

Change-Id: I0f483ea037fc76a362f1d174b521075e8e9b0a3a
Reviewed-by: Sergey Belyashov <Sergey.Belyashov@gmail.com>
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>