Make the enum value descriptions consistent
Currently, they are inconsistent and very hard to read. Hence, the minority is
changed to follow the majority, and also the "table" style used throughout Qt.

I also added a line break where the value explanation is long to have a mode
explicit separate between the values. Otherwise, the flow of the text is too
crowded, and the reader can lost the track.

Change-Id: Ia14162b162357ec2c85be1726b7f220e1e6b6ec6
Reviewed-by: Denis Shienkov <denis.shienkov@gmail.com>
Reviewed-by: Sergey Belyashov <Sergey.Belyashov@gmail.com>